+++ src/assetbundles/src/js/settings-plugin.js
@@ -1,3 +1,6 @@
+import '../css/settings-plugin.css';
+import ClipboardJS from 'clipboard';
+
 var inputField = document.getElementById('webhookEndpoint');
 var inputContainer = inputField.parentElement;
 var copyButton = document.createElement('a');
@@ -7,4 +10,5 @@ copyButton.setAttribute('data-clipboard-target', '#webhookEndpoint');
 
 inputContainer.appendChild(copyButton);
 
-var clipboard = new ClipboardJS('.copy-btn');
+// TODO: visually confirm copy with user
+new ClipboardJS('.copy-btn');