+++ src/models/Settings.php
@@ -66,6 +66,23 @@ class Settings extends Model
     public $notificationEmails = [];
 
     /**
+     * @var string optional path to a custom template to be used for admin order
+     *             notification emails
+     */
+    public $notificationEmailTemplate = '';
+
+    /**
+     * @var bool
+     */
+    public $sendCustomerOrderNotificationEmail = false;
+
+    /**
+     * @var string optional path to a custom template to be used for customer
+     *             order notification emails
+     */
+    public $customerNotificationEmailTemplate = '';
+
+    /**
      * @var array
      */
     public $enabledCurrencies = [ self::CURRENCY_USD ];